Fix accessibility issue (link purpose in context)
The "manage your email subscriptions" page lists all the email
notifications a user is currently subscribed to. Each of the items on
the list has a "Change how often you get emails" link and "Unsubscribe"
link. The links do not specify which topic they are connected to,
therefore to a user navigating the page using a screen reader, it would
not be so clear what the purpose of the links is.

This adds some hidden text to the links which references which
page/topic they are related to. This way, the page will appear the same
as before to sighted users, while improving the experience for screen
reader users.