Remove proof-of-concept account auth
This commit removes the "things are changing" page and the "use your
GOV.UK account" message from the sign in page.  These were set up as a
proof-of-concept only enabled on staging, for us to see how linking
notifications with the account could work.  We're now building the
actual pages for the real user journeys, so we don't need these.

This commit does not remove the "use your GOV.UK account" page which
you see if you try to sign in with linked accounts, as that page is
remaining (though with different content).