Add a confirmation page for creating an account subscription
This page handles signing up to notifications connected to the
account, which means it also handles the "linking" of an account-api
`OidcUser` to an email-alert-api `Subscriber`.

If a user has subscriptions which pre-date their account, we'll list
them so that the user knows what is changing.  The linking process
also triggers an email in this case (email-alert-api does this).

This page is intentionally very generic: it's not just for single-page
notifications, or content item subscriptions, it's for topics in
general.  This is so we can use this page across different sorts of
email journeys.