Add Feature: Single page notification signup logic
Previously users have been able to sign up to notifications about
collections of content, and get updates when new items are added to that
list.

We now want users to be able to montitor for changes / additions to
single items of content.

When a user clicks "Get emails about this page" on a single piece of
guidance, we will need to perform various checks to ensure their account
is ready form them to be subscribed.

Specifically:
Only a logged in user, with valid account session, whose previous
notification account subscriptions have been linked can subscribe to
single page notifications.

When the user is logged in we perform various API calls to automatically
link their accounts and subscribe them.

When the user is not logged in we show an informational page to ensure
they understand what is about to happen before they log in/sign up and
subscribe.

Co-authored-by: Alex Whitehead-Smith <alex.smith@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <huw.diprose@digital.cabinet-office.gov.uk>