Add Development manual test form for single page notifications
The new `/email/subscriptions/single-page/new"` path is what users will
post to from guidance when they click "Subscribe to updates".

As this is a POST request, it can be difficult to see the rendered
informational page to check content for logged out users.

We've added a form to quickly get you there if you are in a development
environment.

Co-authored-by: Alex Whitehead-Smith <alex.smith@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <huw.diprose@digital.cabinet-office.gov.uk>