Implement subscriber account authentication
When a user comes to `/email/manage/authenticate`, they enter an email
address in a form.  We want to then:

- just log them in to email-alert-frontend, if they're logged into a
GOV.UK account and that GOV.UK account has the same email address; or

- tell them to sign in with their account, if there is a GOV.UK
account with the same email address but they're not logged into it; or

- fall back to the current magic link journey if the email address
does not belong to a GOV.UK account.