Get users to sign in to create a topic subscription for an account
When setting up a topic subscription, the previous commit bypasses the
email form if the user is signed in.  If the user is NOT signed in,
and so gets prompted for an email address, we want to check if they
have a GOV.UK account: if they do, we prompt them to sign in to it;
otherwise we fall back to the current magic link journey for approving
the subscription.