Add /sign-in link to inset on interrupt page
The interrupt page is intended to give users more information before
signing up or registering to a linked GOV.UK Account and notification.

We've recently launched the /sign-in page that is intended to help users
navigate through account confusion. We can link to that page here, so
users can explore onwards if they are confused at this point, and return
to this journey later.