Add single page unsubscribe
If a user is already logged into their GOV.UK account and clicks the single
page notifications button on a page, we now check if they're already
subscribed. If they are, unsubscribe them and redirect them back to the
page.

This exists on the same route as the subscriptions so we can do a graceful
fallback if the JavaScript that updates the single page notifications
button on the content page fails for any reason.