Pass topic_id of single-page subscription to account confirm page
The confirmation page expects a `topic_id` parameter, which is the
slug of the `SubscriberList`.

This commit has a slight change of behaviour: previously we wouldn't
call email-alert-api to find-or-create the `SubscriberList` if the
user were logged out.  Now we do, so that we can get the slug.

I also changed some `[]`s into `fetch`es where the values are
required, and removed a redundant `rescue` (which is handled by a
`rescue_from` in the `ApplicationController`).