Do not create a session cookie for account users
If a user is signing in to this app via their GOV.UK account, we don't
want to create a separate session cookie: if we do, we then have to
tie together deleting the email-alert-frontend session cookie when we
log a user out, and it's potentially confusing that the
email-alert-frontend session cookie will expire before the
`__Host-govuk_account_session` one

So, we change the authentication in this app to *preferentially* use
the GOV.UK account (see `authenticated_via_account?` ahead of
`session["authentication"}.present?` in the conditional).

When using account auth, no session cookie is created, we fall back to
linking with an email-alert-api subscriber on every request.  This
will be slower, but we can keep an eye on performance, as we expect a
gradual ramp-up of notifications users also using an account.

This does mean we have two totally separate authentication mechanisms
in the same app, until we migrate all notifications users over to the
GOV.UK account, which is a bit of a code smell.