Show last update date for single page subscriptions
Since these subscriptions are for a single page, we can show the user
the last time they were updated so they can see if and when they would
have received an email about it.

This does call the content store once per single page subscription the
user has which could cause performance problems if a user has many
subscriptions. It should be fine for now but it's something we should
keep an eye on.