Move new subscripiton controller method from :show to :create
This was a bit of a RESTful etiquette error on my part. We're doing a
POST request here, by convention that should hit a :create route.

We made it hit the :show route here, because we'd set it to respond with
a view if the user was not logged in. In the next commit we'll change
that so that we redirect to a GET based :show path instead.

So this commit just does the initial renaming of :show -> :create as a
tidy up ahead of this change.