DRY out topic ID and require it on :edit
The POST route will call fetch_subscriber_list and grab the topic ID
from the content store.

Beyond that the user either:
- gets subscribed and returned to their content.

Or
- passed to :show and given the chance to opt in, with a topic_id param
- upton clicking passed through :edit requring a topic_id param (and
getting analytics added along the way)
- finally redireted to the log in / sign up with all required params.

As Topic ID is set once in the POST request, and is a requirement for
:show and :edit, we'll need to fetch it off the URL and throw an error
if it is not present. We can then catch this to serve a 404 if a user
has meddled with their URL.

Initially i also memoised this, but I changed my mind about that, it
brings the fetch safeguard into question, as sometimes it is stored in
the instance variable, and other times we're getting it off the URL.

I think it also created some flakiness based on the order they are run
in (and so what's memoised). Plus it's hardly a performance saving here.

As a result we do have a bit of strangeness, it's important the
@topic_id set by the fetch_subscriber_list must be used as an instance
variable, as it is arriving through that API cann and not off the URL.

For :show and :edit we're getting it out of the params, so there's a
difference internal to the controller now.