Fix default email frequency for signed out users
In #1193 we changed the default email frequency from 'daily' to
'immediately'. However, we only changed it in the `SinglePageSubscriptions`
controller which meant that users who need to sign in and then return to
`AccountSubscriptionsController#confirm` were still getting the 'daily'
frequency.

Set the frequency parameter in the callback URL so it's passed through
to the other controller. Also take the opportunity to define it as a
variable in the controller to avoid repetition.