Bring black flash message for topic subscriptions
We introduced a regression, where creating a topic subscription
through the account would not show a success flash message on the
manage page.  This commit fixes that, and also updates the content of
the flash to match our designs.

I could have made the template use the account flash, but decided
that, as we have both account and non-account topic subscriptions for
now, it would be best for them to both use the same approach.  So
these flash messages do not use the account flash.

I also changed some places where we were redirecting to the manage
page via the `process_govuk_account_path`, even though we know at this
point the user is signed in.