Handle signed out users who are already subscribed to a page
Previously, if a user was signed out and clicked the 'get emails about
this page' button to subscribe to email alerts about a single page, we
didn't check if they already had a subscription to that page.

If they did, we'd ask them to confirm again and send a new confirmation
email, but the manage emails page would show the date of when they
originally subscribed which was a bit confusing.

Now, do that check and if they're already subscribed redirect them back
to the content page after they sign in with a flash message to display
an info banner telling them they were already subscribed.