When a user is already subscribed, keep analytics params in redirect
For cross-domain tracking with accounts, our rule of thumb is that
analytics parameters (`_ga` and `cookie_consent`) need to be preserved
through all redirects until the user lands on a page which renders an
actual response (and so runs the analytics javascript which acts on
those params).

Since DI can send users to this controller method, they may arrive
with analytics parameters from the other domain.  If we immediately
redirect the user away (rather than rendering a response), we need to
preserve those parameters, or we risk losing tracking.