Show subscription name on the update frequency page
Previously there was no mention of what subscription you're changing the
frequency of on this page which was confusing. Add the title from the
subscriber list so people know what they're changing.

Having `@subscriptions[@subscription_id]["subscriber_list"]["title"]` in
the view is a bit ugly, so add a new `SubscriptionsManagementHelper`
module to hold a new method that is only used in the view.