Pass analytics params when signing in to manage page
When an unauthenticated user tries to go to `/email/manage` they are
redirected through the sign in process. If they sign in with a GOV.UK
account this means they'll be taken off the `www.gov.uk` domain which
breaks the analytics tracking when they return as we weren't preserving
the linker parameters GOV.UK sign in are passing back to the app.

Now, pass through the parameters in the final redirect so they're
present on the page that Google Analytics loads on. Also add a test so
we don't accidentally break this in the future.