Add account privacy information to subscription confirmation page
In [[1]] we removed the first time sign up page that asked users for
consent to use cookies, to email them for feedback, and displayed
privacy information about how we use the data associated with their
account.

We're not going to be collecting consent for cookies or feedback
centrally in the account any more, so that doesn't need moving, but we
do need to show users the privacy information at some point in the
creation journey.

Add this information to the confirmation page instead, so users can
read it before finalising their subscriptions.

[1]: https://github.com/alphagov/frontend/pull/3224