Minor markup adjustment on Manage your emails page
Adjustment to the "Subscriptions for emailaddress" snippet so that is no
longer a heading but a regular paragraph with the email address in bold.

Also, use have_content instead of include in test:

The subscription management controller test is failing because include
doesn't seem to recognise content the same when you throw an html tag in
there. I.e. it sees "Subscriptions for <strong>email</strong> as
different from "Subscriptions for email"