Disable source map line comments for css in production
This removes the line comments for css in production as we don't need
them in production and only increase css filesize. Line comments make
assets differ depending on compilation environment. This is an issue
when we build the application for our EKS environment vs our old as our
assets differ and ultimately end having a different filename hash. This
makes it difficult to do a traffic split as the css will have a
different filename for old vs new infra.