Format params for email-alert-api
Currently this app provides two endpoints for a user signing up
to email subs. The content_items_signup controller handles requests
that bypass the need for a govuk-account. The single_page_subscription
controller handles requests that enforce the govuk account.

We want to allow users to sign up to a document collection ie a single page
of content without sending them down the account route. This spike is one way
we might do that.

I've configured the params for email alert api based on the presence of a paramater.
An alternative method would bave been to check the content_type of the content
item that's being subscribed to. ie if this piece of content is a document collection
then send the content id. But that felt hacky and harder to maintain. This way, if we
want to allow more document types to be subscribed to without an account, we won't need
to make any further changes to this app.