Update account-related content
Update content on the single page subscriptions "interrupt" page aka
"The way you subscribe to get emails from GOV.UK is changing" page.

This update is in order to bring this page in line with a series of
updates the accounts team are implementing on the GOV.UK account. It
also removes the claim that you need a UK phone number to create an
account as this is no longer a requirement.