Hotfix for single page subscriptions
This code was originally introduced to handle special cases for subscriptions. However, this affected sign-ups for other pages using this same flow, since the value of the `single_page_subscription` can be present but a string with the value of false. This meant that for non single page subscriptions, the subscriptions were still going through the single page flow, resulting in users not receiving email alerts for content changes.