Improve controller tests
Explicitly test that the correct params are being sent to email-alert-api.
If the single_page_subscription paramater is present, and set to "true",
params are formatted by the GenerateSinglePageListParamsService, if it is not present,
or is set to anything other than "true", they are formatted by the GenerateLinksBasedListParamsService.

For reference, this parameter is being sent by the frontend component
single_page_notification_button [1]

The gds-api-adapter helper "stub_email_alert_api_creates_subscriber_list"[2]
was previously used in this spec, but that doesn't allow us to check that
the correct params are being passed to the create action.
[1] = https://github.com/alphagov/govuk_publishing_components/blob/1c74d79419bf56e640c3dc3a372e3a73fd8eef5c/app/views/govuk_publishing_components/components/_single_page_notification_button.html.erb#L17
[2] = https://github.com/alphagov/gds-api-adapters/blob/main/lib/gds_api/test_helpers/email_alert_api.rb#L127-L133