Refactor hotfix
Cause of bug:
The param `single_page_subscription` can be sent to the content_items_signups_controller
from a frontend component[1]. When present, the content_items_signups_controller will format
the subscriber_list params it sends to email-alert-api to build a content id based
subscriber list (aka a single page subscription) rather than the default links based subscriber list.

- The `new` action handles the incoming request and`params[:single_page_subscription].present?` was used to
return true or false.
- The `create` action POSTs the params to email-alert-api. So it was necessary to pass this boolean via a hidden_field_tag
in the confirm template so that it was available to the create POST request.

Rails translates the paramaters passed into the hidden_field_tag to HTML, so the bool false is turned into the string
"false"[2].This was the cause of the bug. Because now when the create action runs the code params[:single_page_subscription].present?
it will return true, instead of false, and the params will be formatted for email-alert-api incorrectly

Changes in this commit:
Instead of passing the paramater from the new action to the create action for all requests,
and using "true" or "false" to decide how the params are formatted, ONLY add the hidden_text_field
if the paramater is present.

This doesn't change behaviour - ie the hotfix worked fine - but it's cleaner and should have been done in the
first place.

[1] https://github.com/alphagov/govuk_publishing_components/blob/main/lib/govuk_publishing_components/presenters/single_page_notification_button_helper.rb#L63-L77
[2] https://stackoverflow.com/a/50930214